+++ Temp/StagingArea/Data/il2cppOutput/t858.h
@@ -4,5 +4,5 @@
 #include "t858.h"
 struct  t858 
 {
-	int32_t f1;
+	uint8_t f1;
 };
